Fix dark mode for chartViewport.css
This updates perf_ui/chartViewport.css with using the declared color
variables. It also introduces missing colors for selecting a range
of a portion, which is used when shift-dragging a portion of the
performance panel timeline, as well as the colors for (Mac) scrollbars.

R=petermueller@chromium.org,jacktfranklin@chromium.org

Bug: 1152736
Change-Id: I04a9759aa92caf81f67c9f3d3e0ff3dfb2deec2e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2721956
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>