Fix dark mode for perf_ui/flameChart.css
For hovering/selecting a node in the flame chart tree, we add a new
CSS variable to add some slight gradient. It also changes the
selection color of a node in the flamechart from blue to red (e.g.
the change from --accent-color to --accent-color-red).

R=jacktfranklin@chromium.org,petermueller@chromium.org

Bug: 1152736
Change-Id: I4a5e0da38dafd485adfedd1ed3a786e6ad62007c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2722440
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>