Translate issue category names on demand
The current version translates the strings at module instantiation
time. This is discouraged as translated strings are most likely not
yet available.

This CL changes the category view to translate the category name
on-demand. The lookup is fast enough so caching the name is not
really necessary.

R=sigurds@chromium.org

Bug: chromium:1136655, chromium:1180760
Change-Id: I4b6baf9ffe8278defa95b547060e1a69bbbcffef
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2726571
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>