Add helper functions for adding breakpoints in sources
In this CL we add and improve the helper functions that
allow us to test the breakpoint scenarios. In order to
do so we have improved the waiting conditions for loading
a file in the sources panel by adding an event that fires
when the file finishes loading or reloading and listening
to that event in the tests.

Bug: chromium:1169599, chromium:1112692, chromium:1178804
Change-Id: I315c722c0e5efdea747ba55b729995f5aee7376c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2698249
Commit-Queue: Almothana Athamneh <almuthanna@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>