Fix DragHandler for usage in web components
The DragHandler is not working correctly below shadow roots.
This CL changes that by attaching the mouseout handler to
the getRootNode() instead of the document.

This caused flickering when resizing columns in the
DeveloperResourceView and the Coverage view. Other
parts of DevTools might have been affected as well.

Bug: chromium:1174408
Change-Id: I186653822b521f01066cad36e7b294cfc33eca4a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2731908
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>