[eslint] Fix bug in rule that looks at every top-level object literal
The rule looks up the 'name' attribute on the 'VariableDeclarator'
node, which does not exist, instead of the 'name' attribute on the
corresponding Identifier node.

This CL fixes that by extracting the Identifier check into a common
helper and adding a regression test.

R=petermarshall@chromium.org

Bug: chromium:1180760
Change-Id: If82b24fb27f12a4ebd665b99daf5dea747bc866b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2734617
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>