Use relative URLs for all images
Now all images in DevTools use relative file lookups. In the
Icon component, we construct the URL inline and pass it on to
the styles computation. Note that we also wrap the render function
in a coordinator write block, as the render function will cause
DOM updates to be scheduled.

R=aerotwist@chromium.org,szuend@chromium.org

Fixed: 1183361
Change-Id: Ieb23a0a7d059cbbb408abc4ced9398ea1b0d44e8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2732381
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>