Allow Esc key to dismiss tooltips without toggling drawer
This CL allows the Escape key to dismiss tooltips without toggling the
DevTools drawer. This only affects Escape key behavior when a tooltip
is present; once the tooltip is dismissed, the Escape key will
toggle the drawer as normal.

Currently, pressing the Escape key while a tooltip is present causes
the drawer to toggle, which is annoying if the user is
also using a drawer tool at the same time.

WCAG 1.4.13 requires there to be a convenient way to dismiss tooltips
and suggests using the Escape key. DevTools technically allows any
key to dismiss tooltips, but this is inconsistent with soft context
menus and select dropdowns, which are both dismissable using the
Escape key without toggling the drawer.

The drawer toggle behavior when Escape was pressed to dismiss a
tooltip was reported to us during an accessibilty review of the
DevTools.

Before: https://i.imgur.com/YjHXM47.gif

After: https://i.imgur.com/NO3k9U1.gif

Bug: 1184952
Change-Id: I261248ec2696db5ed31a3ea43ff193dbfcbd27b6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2737660
Reviewed-by: Jose Leal <joselea@microsoft.com>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Peter Müller <petermueller@chromium.org>
Commit-Queue: Brian Cui <brcui@microsoft.com>