[LMI] Save value interpreter settings
This adds the feature to save the value interpreter settings across
sessions, consisting of the 1. value types to show,
2. the value type modes for each value type, and the 3. endianness.

Value interpreter settings only make sense as a whole, i.e. if we have
several LMI views open, we do not want to save settings that result
from individual changes of several LMI views (e.g. value type from
one view, and endianness from the other view).

This CL triggers a SettingsChangedEvent every time
a user changes the settings for one view. This results in the
LinearMemoryController saving the three mentioned items that are used
if a new view is opened. Individual LMI views are still allowed to
use separate settings, as the saved settings are only applied
whenever a new view is opened for the first time.

Drive-by: Use LinearMemoryInspectorModule for clarity in imports
Fixed: chromium:1182142
Change-Id: I702b9a3350ebc6bfb9f138815c9dd78ecd41c92b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2731569
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>