Speculatively increase timeout in before hook of e2e test
The before hook of test/e2e/inline_editor/font-editor_test.ts
is timing out on bots. This might be a problem with a waitFor
in the hook, or the timeout might be too low. This CL speculatively
doubles the timeout so we can see whether it is due to the timeout.

Bug: chromium:1184658
Change-Id: Ie7696190f087dc57e9b263c5cae96827ea8cf433
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2744584
Auto-Submit: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>