Add Vary HTTP response header column in Cache Storage viewer
This CL adds a new "Vary Header" column to the Cache Storage viewer to
help web developers that fail to match or delete a cache entry request.
It happens when `ignoreVary` is not set to true. A tooltip is now shown
when the Vary header is not empty.

Bug: 1186049
No-Presubmit: true
Change-Id: I2dd7f926ba8f81861a697cc327fa0420f747fc64
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2748082
Commit-Queue: François Beaufort <beaufort.francois@gmail.com>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>