[e2e] Cookies tests: Reenable and wait for the Cookies item's children
Failing tests all windows fail because the double click on the Cookies
item in the application panel doesn't work. One reason could be that the
item is clicked before it has any children. This CL explicitly waits for
the children marker to appear before attemptying to click.

Bug: chromium:1186150
Change-Id: Idafd9c6d669ab554fa352168bf64f4204c8c9a68
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2760240
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>