Allow shebang in ESLint license header check
Noticed when migrating a script to JS; we don't currently support
allowing a shebang at the top of a script, it gets put below the license
comment which is invalid.

This CL updates the rule to allow a shebang at the top of the file.

Bug: 1186163
Change-Id: I64a13b086452f3b7da93fa7c3dcd85521463cda7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2760242
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>