Explicitly focus BreakpointEditDialog editor after the dialog is shown
This dialog is shown when a breakpoint is requested to be edited in the
sources panel. For this, it utilizes an instance of
CodeMirrorTextEditorFactory which was previously instanced
asynchronously via legacy runtime extensions. This async instancing
apparently introduced a race condition which allowed the editor inside
the dialog to be focused via a call inside the constructor of
BreakpointEditDialog. However, when migrating @UI.TextEditorFactory
extensions we moved to synchronously instancing
CodeMirrorTextEditorFactory, which affected the race condition and
resulted in the editor not being focused automatically when the dialog
is shown.

To make sure we focus the editor we add a call to do so after the
dialog is shown.

Also, some type annotations are refactored to be more explicit as a
allowed by the migration.

Fixed: 1187845
Change-Id: Ibe3331c537cd470c48be7fc308639fc77bd39319
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2761000
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>