[LMI] Add 32/64 bit pointers in ValueInterpreter
This adds 32/64 bit pointers to the ValueInterpreter, including
the 1) ValueInterpreterDisplay and the 2) ValueInterpreterSettings.
With this change, the user can now also interpret values in the
LMI as pointers.

The 32-bit pointer interpretation also has an additional ¨go-to-link¨
button next to which allows the user to jump to that
specific address in the LMI. This is not yet implemented for the
64-bit pointer (more changes are required for using big ints in
the LMI).

Screenshot: https://imgur.com/a/0N3T6LA

Bug: chromium:1184436
Change-Id: Ib5c52d0eb8cc4386775633b1ed49a7001f5158de
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2763867
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>