Apply source maps to all outputs that look like an Error stack
Previously, we'd only try to apply source maps to the following
error names:
  EvalError, ReferenceError, SyntaxError,
  TypeError, RangeError, Error, URIError

This CL applies it to everything ending in Error. This is a
developer pain point, and we're landing this band-aid until
we have a way to report structured stack traces from V8.

Bug: b/173826328
Change-Id: Ie6a0465c64ab3b442ac5359517e800dcacb88cb0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2766902
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>