Wait for networkidle2 event when reloading DevTools in E2E tests
To load DevTools in E2E tests, we do something like this:

await frontend.goto(url, {
  waitUntil: ['networkidle2', 'domcontentloaded']
});

However, when reloading devtools (which we do after enabling an
experiment for example), we do something like this instead:

await frontend.goto(url, {
  waitUntil: ['domcontentloaded']
});

I found that this could be a cause for race conditions. I was working on
a test (not landed) where I was trying to open QuickOpen immediately
after having reloaded devtools, and nothing was happening.
Adding the waitUntil on networkidle2 fixed it for me.

Beyond that, I think it's better to be consistent on how we load and
reload devtools.

Bug: 1171673
Change-Id: I48a00518ceccc86df8249e7728d064acc80ab930
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2767421
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>