Walk up TreeOutline containers to find one that scrolls in _deferredScrollIntoView
This fixes a regression where the Elements pane wasn't scrolling anymore
while hovering in the page with the inspect tool.

In the past, the way we scrolled TreeOutline items into view was by
simply using el.scrollIntoViewIfNeeded.
This worked well and was simple, but had one limitation: it always
scrolled along the 2 axes even if the element was already visible along
one of the 2.
This led to user frustration when, e.g., using the Sources panel as the
file tree would sometimes be slightly scrolled to the right instead of
always aligned left (see crbug.com/1040223).

The way this was fixed was by removing scrollIntoViewIfNeeded, and
instead calculating the minimum delta required to show the item in the
view and calling scrollableContainer.scroll(detla).

This unfortunately didn't account for the Elements panel. In the
Sources panel, the TreeOutline's shadowRoot container is the one that
scrolls. But in the Elements panel, the scrollable container is one
level up in the DOM.

So the fix now is to make sure we get the right scrollable container
into account before calculating the delta and doing the scrolling.

Before/after comparison: https://imgur.com/a/RnnCF8s

Bug: 1149842
Change-Id: I6220e66d39869772fa4cb42fc04b5229ecc46b6e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2765163
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>