Fix broken context menu actions in Sources text editor
The context menu of the Sources text editor had been broken as a
regression of https://crrev.com/c/2550688 and it was subsequently fixed
in https://crrev.com/c/2567957 . However, the fix in the latter altered
the original behaviour by relying on CodeMirror's `on` API to listen for
`contextmenu` events instead of the native `addEventListener` API as
originally implemented. This transitively provoked a change in the
flavors added to the application context by the time the actions are
executed, causing them not to run successfully.

This CL restores the original behaviour of the actions in the context
menu by adding the listener via the `addListener` API in the
SourcesTextEditor element.

Fixed: 1189118
Change-Id: I85cfafb273856cd5ca9ee53500e7b0af26ee9d11
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2770578
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>