Add distinct shortcut for triggering Sources editors action on mac
These actions don't have a dedicated shortcut for mac and aren't
triggered on this plaform using the currently bound shortcuts as they
use the `Ctrl` key (this key is generally mapped to the `Meta` key in
mac). This CL adds a dedicated shortcut for the mac platform to each of
these actions using the `Meta` key.

Bug: none
Change-Id: Ib060a669c7c9018c5c5462ca0e60017b977262d0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2770579
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>