Add 'Show More' button to stack traces in frame details view
`JSPresentationUtils.buildStackTraceRows()` includes a logic to not
show all rows of a stack trace. Stack trace frames can be set to hidden
either because they are in source files which are on the ignore list, or
because the stack trace already contains a lot of rows.

This CL implements the logic to actually hide those rows in stack traces
in the frame details view and adds a button for showing all rows.

Screenshot: https://i.imgur.com/N73fTr2.mp4

Bug: 1159334
Change-Id: Id6fe171b5ab4bbb3136599e5c0bac5b5ab21bde3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2764408
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>