Adds unit tests to the What's New panel
This way we remove the dependency of the panel on global variables,
which is breaking it because they are no longer initialized since the
removal of the help-legacy.ts file from help's module.json.

Also includes the following changes to the unit tests infra:
deinitializing global vars in EnvironmentHelpers now removes the
ViewManager instance and resets the registered settings. These, as
needed by the unit tests of the What's new panel to work properly.

This needs https://crrev.com/c/2773409 to land first

Fixed: 1190068
Change-Id: I4ebe643f9143bf86ffa45df449ba5e067edb0916
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2774676
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>