Update e2e mocharc to read files from new test runner
This CL updates the e2e tests so that they read config from the new
test runner, should they be being run by the new test runner. Note
that currently we do not use the new test runner in any of the `npm
run` scripts or on the bots, but this CL is the first step towards
doing that.

Bug: chromium:1186163
Change-Id: Ia7885c9e47145ffa686ed501cdbb5af4818829c5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2778318
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>