[LMI] Only record addresses once if same address is selected
When the LMI is resizing it may require more data to be fetched for
showing it in the view. This will kick off a chain of events, which in
the end leads to the LMI getting new data. This new data also contains
the address field. The address will have the same value as before, as
the only change is the update in the data that is provided for
rendering. If the same address is once again set on the LMI, we do not
want to record it in the history, as we'd otherwise have duplicate
entries which lead to confusing history behavior.

Bug: chromium:1166577
Change-Id: Ie553f2c5cb3ab66fd03350c642b38d55141f7ed6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2778594
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>