Do not include TypeScript build info in data_deps
We should not include all TypeScript build output in our data
attributes. We are using `devtools_module` et al to make sure
we always include the correct information, but we can't use
that for our legacy layout test runner infrastructure, as they
should not be included in our main bundle.

Normally this should not be a problem, as the .tsbuildinfo
should have been made deterministic when we previously discovered
and reported this issue in
https://github.com/microsoft/TypeScript/issues/37156

It seems like the TypeScript compiler has regressed in terms of
non-determinism, which we will be reporting to the TypeScript
team.

R=jacktfranklin@chromium.org

Bug: 1190355
Change-Id: I2ea8b09a93e25762843fab8e0e4aef25fa5fac62
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2797197
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>