Allow vertical toolbar items to wrap and don't clip them
Some tools use vertical toolbars (network and console). One problem with
this type of toolbars is that items are set to be 26px high and are not
allowed to wrap. When space starts to become tiny the items get
truncated which is a problem for accessibility.

This often happens when you zoom, or have a narrow devtools (which is
often the case when docked to the side).

Horizontal toolbars are not impacted in the same way because they are
capable of wrapping items on multiple lines.
Vertical toolbars however are already on multiple lines. So the fix here
is to allow individual items to have their text wrap.

Before/After screenshots: https://imgur.com/a/uodiw45

Bug: 1194986
Change-Id: I4f7d9bc7dca24c394f0e9d956c7866e4a753d570
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2799434
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Songtao Xia <soxia@microsoft.com>
Reviewed-by: Changhao Han <changhaohan@chromium.org>