Move QuickOpen.FilteredListWidget.Provider extensions to sources-meta.ts
These extensions register the providers that fill the quick open menu
with content from the sources panel. To avoid statically importing
quick_open inside sources-meta.ts we opted not to register them inside
the meta file but in the module's implementation. However, this in turn
means that the extensions won't be available until the `sources` module
has been lazy-loaded.

After taking a deeper look, it was found out that the quick_open module
is loaded at bootup anyways, as it is a transitive dependency of the
`main` module. This means that registering these extensions inside
sources-meta.ts and, thus, statically importing the quick_open menu
wouldn't add an extra overhead to the loading time.

So that these extensions are available from bootup, we move their
registration to the module's meta file.

Fixed: 1195352
Change-Id: I7f0f8da25d021c10dc70894ff0b69f414a0b409f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2803968
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>