[A11yTree] Hook up listeners for DOM overlay/inspector
Adds event listeners for selection/mouse over events in nodes of the
tree to show the DOM overlay and set the inspected node in the DOM tree.

Note: When swapping back to the DOM tree, it always scrolls back to the
top of the tree. It should probably either:

a) Maintain scroll position when swapping between a11y and DOM trees
b) Scroll the inspected node into view.

Bug: 887173
Change-Id: Id89b69366ba69a5d606b05ca834b6ec46c70f091
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2783093
Commit-Queue: Meredith Lane <meredithl@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>