[TestRunner] allow e2e resource path to not start with a slash
This is for consistency with every other path option that the runner
takes; because we `path.join()` them they should work with or without
a starting slash. This is the only flag that requires a slash, so I'd
like to change it to make it work with or without.

Bug: chromium:1186163
Change-Id: I2f55da0a82b0d80d33f4d1c8f25c85492cdefd25
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2810279
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>