[ARIAUtils]: Remove element parameter for alert()
This CL removes the element parameter from ARIAUtils.alert.  The element parameter is used to access the owner document.  In place of that, this CL introduces ARIAUtils.setOwnerDocument which is called in MainImpl._showAppUI.  This change allows ARIAUtils.alert to be called when an element is not readily accessible.

Bug: 1195232
Change-Id: I646b604d281ad32bd0811514a3bfbc3254f998c4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2801093
Commit-Queue: Michael Liao <michael.liao@microsoft.com>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Kalon Hinds <kahinds@microsoft.com>
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>