Bundle meta files into separate entrypoint bundles
By bundling all meta files (which happens automatically, as Rollup
only treats filenames that are equivalent to the direct folder
they are in, which `-meta.ts` files by definition do not), we
reduce the amount of network requests performed at startup.

Before this CL:

Mean boot time: 1115.99ms
50th percentile boot time: 1115.28ms
90th percentile boot time: 1122.66ms
99th percentile boot time: 1159.44ms

After this CL:

Mean boot time: 1105.17ms
50th percentile boot time: 1072.34ms
90th percentile boot time: 1111.8ms
99th percentile boot time: 1442.42ms

Note that this is a marginal difference, but run on good hardware.
The theory is that this change will be more benificial for users
running slower hardware.

R=jacktfranklin@chromium.org

Bug: none
Change-Id: I24aa02d9f162d1ac07ceae5df2beaa0a61a7dd43
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2790878
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>