Remove dagre_layout
This code is unused and will not be shipping with DevTools in the
near future. It is also third_party code, so we will need to follow
the third_party procedure if we want to add back this code.

R=aerotwist@chromium.org

Bug: 1187573
Change-Id: I83af5983bf5223dc167ac2d9c824b39a559df717
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2821875
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>