Improve color contrast in the box model diagram
This changes ensures that borders between the various box-model regions
have enough contrast compared to the region background colors so that
low vision or color deficient users can still easily understand the
diagram.

The colors of the regions don't depend on the selected theme since they
come from the overlay colors. So we now use new colors for the borders
that don't come from the theme either.

This change also improves the hover behavior in dark theme.
On hover, only the hovered region has a color. The other ones become
transparent which means the panel background is visible. In light theme
it's fine because the borders and text are already dark and threfore
remain fully visible.
In dark theme however, we need to switch the colors of the non-hovered
regions to something lighter. The code to do this is a little bit
complicated because regions are nested into eachother and colors set on
a parent container get inherited by children.

Before/after and hover screenshots: https://imgur.com/a/ySTLfGj

Bug: 1197542
Change-Id: Iac6b549a847ae4f6bf1eebe96c771109e792ed41
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2817904
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Kalon Hinds <kahinds@microsoft.com>