Re-filter console messages when a message's live location changes
Filters for console messages can already be in place when a page is
freshly loaded. Should a message match a filter, but then due to
a change to the message's live location later no longer match that
filter, the console message is still shown/hidden.

This CL fixes this bug by re-filtering console messages when
message's live locations change.

Please note that this can cause a one time flicker of console
messages. This is due to the inherent timing difference between
when a console message is first rendered and when the live location
later updates because of script parsed events or source map
loading.

R=wolfi@chromium.org

Bug: chromium:1166269
Change-Id: Idc59599ff1a0b5811c61c5315927fdec76e5d361
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2825428
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>