Rewrite test after SABs have been disabled by default
The SAB deprecation test is no longer working, as in non-COI contexts
the SAB constructor is not available. This CL rewrites the test to use
a cross-origin COI iframe, in which the SAB constructor is present,
but SABs can't be transferred (which is simlar to what we had in
non-COI main frames before SABs got disabled by default).

Bug: chromium:1196618
Change-Id: I07b9c3acd5e4297d6efd0052784628dc62ad5e32
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2828293
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>