Improve type safety by allowing issue code to be of enum type
This CL makes Issue a generic class that takes a generic type for
the issue code. This type constrains the return type of Issue.code()
such that callers that, e.g. are dealing with a CorsIssue know that
the return values are coming from a particular enum value.

This allows to easily check that displaying code handles all codes.

One caveat is that the types cannot guarantee for IssueAggregator
that the aggregated issues are always of the same type although
this is the case. The reason is that this would require global
knowledge (we would have to know on the type level that no two issue
codes in subclasses of issues are the same).

Bug: chromium:1072335
Change-Id: I741e50db882636de314b79cd2109e2cb6efcf6e7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2848237
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>