[DarkMode] migrate ConsoleView.css
I was going to generate a CSS file for this but then I realised that
it wasn't so many that I couldn't do it by hand. Along the way I found
a few CSS selectors that I'm pretty sure are unused; I couldn't find
them when testing with a variety of console inputs, so I've removed
them. They also relied on CSS classes that I couldn't find references
to in `ConsoleView.ts`.

I've made liberal use of `--override-*` variables to get this file
migrated but will sync with Peter as I think some of these may be
legitimate color variables to pull into the theme colors.

Bug: chromium:1152736
Change-Id: I6bf5cab47d37b86bac5d820cf89c80d29bb999b9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2853546
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>