Don't add model observers multiple times in ApplicationPanelSidebar
The ApplicationPanelSidebar is adding model observers on each call
to initialize(), which is accidentally fine at the moment, but breaks
if the CachedResourcesLoaded event is fired multiple times.

This CL moves hooking up the listeners to the constructor.

Bug: chromium:1192177
Change-Id: I4fe7c70749d0d80c8588f1599c0d39c25b3d5a05
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2854684
Auto-Submit: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>