Refactor Settings to make removal internal-only
Removing a setting should only be performed inside of Settings.
Therefore, remove the `.remove()` method on `Setting<V>`. This
allows us to remove the dependency on `Settings` as well.

There was only 1 usage outside of `Settings.ts`, when emulating
a location. If you would remove a setting, it would reset to use
the default value. Since the default value for the location is
the empty string, we should instead call `.set('')`.

R=szuend@chromium.org

Bug: none
Change-Id: I40c17eab5edb968285b18bc51bdf6f558aa05c08
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2859915
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>