Fix the docking icon
The docking icon recently regressed. It was getting overly optimized by
SVGO and was not displaying correctly.
Rather than understanding which setting of SGO in particular is causing
this issue, and turning that setting off (which we cannot do easily),
I'm changing the path code of the icon to avoid the problem.

Bug: 1204985
Change-Id: I2d108b10b466f02fb523368e61577f231c3030b3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2867248
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>