Add wasm resource type to network panel
Adds a new resource type and category for wasm requests.
The mime-type from the response headers is used to change the resource
type of a request once the response is received. This allows requests
to be filtered by resource-type:wasm as well as through the new wasm
filter button.

Fixed: chromium:1103638
Change-Id: Ie307568c5d1fee82a61c1f5e929a4d4b2dea5fdb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2862888
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>