Run PRESUBMIT with Python 3
By specifying USE_PYTHON3 = True at the top of the file,
we run the presubmit checks with Python 3. Most of the
required fixes are related to handling the subprocess
output (which changed from str to bytes) and handling
`len` on the output of `filter`.

R=dpranke@google.com,jacktfranklin@chromium.org

Bug: none
Change-Id: I8740b2a035cf7015cfb972c9b147bd53d49f1bd6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2886934
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>