Network: Fix flickering document previews.
This CL fixes a bug that caused flickering when previewing documents
in the Preview pane in the network tab when clicking documents with the
preview pane already open. This bug was caused by 2 listeners
firing the code to re-render the preview pane in quick succession: one
from selecting a new grid node, and another from activating the
sidepane.

This CL updates the showRequestPane logic in the network panel to early
return if the request sidepane is already open, and we are not specifying
the sidepane tab when activating.

Bug: 1161690
Change-Id: Iabd86029c4c7fb636e2c93600534e0c272ee08ce
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2891083
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Patrick Brosset <patrick.brosset@microsoft.com>
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>