Fix bezier editor showing for custom-property linear-gradients
This issue is caused by an imperfect regex parsing for potential
animation-function-related keywords. Although "linear" is a keyword,
when it is suffixed by "-gradient", it is not a valid animation
function.

This is another example that using regex as a CSS syntax parser will
cause problems. For now this particular problem is fixed, but there
might be other cases that this regex will be false positive or negative.

Fixed: chromium:1203639
Change-Id: I81330260f0470693047c23f7eeb6c86639d5a125
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2896892
Auto-Submit: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>