Fix screen reader issues on DOM Breakpoints checkboxes
When navigating to DOM Breakpoint checkboxes,
the screen reader is announcing inappropriate information
on navigating to checkboxes under 'DOM Breakpoints.
but it is instead should be announcing the title of the checkbox and
whether it's checked or not..
This is caused by by passing a function as localized string .
The CL fixes the issue by resolving the function to a variable.

Bug: 1210160
Change-Id: Id9ea78f3f75ba60e045c1feeaa7de73e78b9f830
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2900828
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jan Scheffler <janscheffler@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>