Make 'Store as global variable' option available for console objects
The option is added by a context menu provider to targets that are
instances of `SDK.RemoteObject`, among other context types (classes)
defined in the provider's registration. However it was not being added to
objects displayed in the console because (1) the target passed to the
context menu is not an instace of `SDK.RemoteObject` and (2) because the
context types of the provider were available only when the sources module
was lailzy loaded (which meant the provider would not be used if the
sources module had not been loaded).

This fixes the bug by passing an instance of `SDK.RemoteObject` to the
context menu triggered on console objects (1) and by making the statically
loaded context types of the provider always available (2).

Fixed: 1209613
Change-Id: I36cf83ff5acf133659c9cbf673e552df0ebdc35d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2905972
Commit-Queue: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>