Add issue for usage of user agent in JS
This adds an issue about the usage of navigator.userAgent (and friends)
to the issues tab. This issue will be triggered on a lot of pages,
but since the change is risky, discussion revealed that this was a
good way of making people aware of the change.

Bug: chromium:1210842
Change-Id: Ib9912d33bf9391772488da1e00fb7760931aee0e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2905504
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Mike Taylor <miketaylr@chromium.org>