Add deadlock detection to render coordinator
It is currently possible to arrive in a deadlock situation with the
render coordinator if you await a read and, inside that, you await a
write (and vice-versa).

This CL adds a race for the reader / writer Promises such that after a
100ms timeout we reject all handlers, and then the dev knows it's a
possible deadlock situation.

R=jacktfranklin@chromium.org

Bug: 1210897
Change-Id: Ibe9b67ded4d93befd2213fb0bc12ab843f9c454e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2906058
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>