[LMI] Fix test to remove check for ArrayBuffers
Previously RemoteObject contained a type check looking for
ArrayBuffers. This was only required because of the set up of a test.
This patch fixes this, such that we do not need to check for
ArrayBuffer types in RemoteObject, as this can otherwise not happen.

Also-by: bmeurer@chromium.org

Fixed: chromium:1196247
Change-Id: I08b2c64bbe194e47b6a8c618485c9f3587a0f297
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2909173
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>