Adding role=group to InfoBars that have actions associated with them
On Windows, with the Narrator screen reader, the "pretty print" infobar
displayed in the Sources panel is announced as a group. This is good
because it's a section of devtools that contains a set of UI elements
but not a major part of the interface hierarchy.
However on Linux, with the Orca screen reader, this infobar isn't
announced as a group.

This CL adds the role=group explicitly to infobars that have actions.

Bug: 1211338
Change-Id: I308568d52c57c706d5ae0cc07245c6c9deb08267
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2910453
Reviewed-by: Songtao Xia <soxia@microsoft.com>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>