Select issues when revealing them
The IssueRevealer scrolls and expands issues in the panel, but because
it does not select them, they don't get focused or announced by screen
readers.

Getting focus on reveal is important for keyboard users who might not
rely only on the fact that the issue got revealed visually, but actually
need it to also be focused.

Demo gif with Narrator: https://imgur.com/Tqs0C24

Bug: 1211749
Change-Id: Ib14b8d22c5517d0558d404ba2c755fac1bededf4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2912353
Reviewed-by: Brandon Walderman <brwalder@microsoft.com>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>