[DataGrid] fix cell focusing on first row bug
This CL fixes a bug where if the user clicked on the first cell of the
first body row, we wouldn't mark that row as selected. This is because
behind the scenes that cell is set to be the cell that we focus should
the user tab into the data-grid. Therefore when they clicked it, the
code that ran considered that cell to be active already, thus meaning
we don't need to re-render to activate the row highlight.

This CL changes the logic slightly to track if the user clicked into
the data grid having not been in it, we will then re-render to
highlight the current row.

This CL has also highlighted that some of the data structures in the
DataGrid could do with a refactor to make this logic cleaner - I'll be
following up with some more CLs next week.

Bug: 1211346
Change-Id: I0657f1ff0fc42470bddb789fce9ca0508ca1870a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2912107
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>