[recorder] Record and replay initial viewport
This CL adds ability to record the initial viewport as the first step
of the recording along with the ability to set the viewport during the
replay. Currently, the viewport would not be reset after the replay
is over. In the future, we should consider if it'd be desired.

Bug: chromium:1205385
Change-Id: Ie69c26163c18dcc891a365e1761a97bd94a3878e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2927304
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Jan Scheffler <janscheffler@chromium.org>