Add aria-label to tag element instead of tag name
We currently rely on ``name from content`` accessible name computation
to assign names to tag elements.
Unfortunately, this can introduce unintended whitespace when the name
computation determines that this is needed for readability.
This change instead labels tag elements with their text content.
This ensures that the name is consistent with the actual markup.

Bug: 1093516
Change-Id: I7e33b0dd349b8b8dc0b960cafb9cee5d4d481bd0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2935283
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Johan Bay <jobay@chromium.org>