[Dark Mode] update legacy color inverted docs
This CL updates the legacy color inversion tool that we have in the
component docs to not rely on the theme patching but instead call the
functions directly. This might seem a little odd, but the goal of the
work is to remove the patching from the main codebase. However for a
period of time it may be useful to generate the dark mode colors as
the legacy patching would have, so I've copied the logic into this doc
so we can use it as required.

Once the dark mode work is done, this can all go :)

Bug: chromium:1152736
Change-Id: Ibce7e35527ade813ce92b7856188812ed3bb18fa
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2940842
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>