[Dark Mode] provide utils for getting color values in JS
This CL adds `getThemeColorValue` as a way to read the current value
of a given CSS variable - the main use case here being to take theme
colors and use them in a `canvas`. It also updates one usage of
`patchColorText`. Future CLs will replace `patchColorText` (which we
are aiming to remove) with this.

Bug: chromium:1152736
Change-Id: Ia8ec00d8a01e5e2e376aa12fe84f71eb97444484
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2940845
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>