Prevent scroll bars from covering dataGrid content on Mac
The old dataGrid component contains elements named corners which add
empty space on the right end of grid rows so that the scrollbar does not
cover actual content. Previously these corners were not shown on Mac. I
assume that is because Mac's default(?) behaviour is to not show
scrollbars unless the user is currently scrolling. But when this is
turned off, the scrollbars can cover content on Mac.

This CL ensures that corners are shown on Mac, so that all platforms
follow the same behaviour.

Before: https://i.imgur.com/mRMdNni.png
After: https://i.imgur.com/DMsFnGb.png

Bug: 1083854
Change-Id: I4d3cce5df18856fce1776994f70bef3320f0992d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2950766
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>