Ensure Network's UIRequestLocation can always be revealed
Previously, revealing a IURequestLocation could only be revealed
using Common.Revealer.reveal if the network panel was already loaded.

This CL factors out the UIRequestLocation to panels/network/forward
so it can be referenced by both network-meta.ts and the rest of the
network panel. This allows the revealer logic to always do an
instanceof check for UIRequestLocation (even if the network module
is not loaded).

Bug: chromium:1196676
Change-Id: I3b9e180bd74e31299a2267268cab5e4ba2f0007d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2953262
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>