Return promise on DOM.removeNode wrapper
The function wrapping the call to this CDP method in the DevTools
frontend uses a parameter callback to call at the method's completion.
However this approach doesn't allow the caller to handle the underlying
promise, for instance, to stop the execution in case of an error.

To that effect, this CL changes the method to return the promise that
results after the completion of the method invocation.

Bug: none
Change-Id: I52a54ee47130b5796b65a167ec1a6b6211f735b6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2957194
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>