Check the element's visibility before showing the tooltip
Up until now, the Tooltip code was using element.offsetParent as a
shortcut to assess if an element was rendered in order to not show its
tooltip if the element or one of its ancenstors is not visible.

The problem with this shortcut is that it fails for elements that are
position:fixed since those also don't have an offsetParent.

Recently, the close button was made display:fixed (to avoid disapearing
when the window was too narrow).
So the idea here is to use the computed style of the element instead.

Bug: 1218816
Change-Id: I7f0831ff62238b3c25462b4d58003afc00d3bff5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2959020
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Leo Lee <leolee@microsoft.com>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>