Add component for linking to network requests
This CL adds a component that can be used to link to a network request.
The component optionally uses a RequestResolver to resolve the request
if only a requestId (in the form of an affected request) is provided.
The component updates its state once the request becomes available,
thereby supporting plasticity of the UI.

Bug: chromium:1217975
Change-Id: I94faa598bb520d4cbdf003ca7858ec759fcc8a6c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2960617
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>