Added new .css.js files to grd_files_debug_sources
We got alerted that DevTools was not loading in chrome due to .css.js
files not being listed in the grd_files_debug_sources. This CL has the
fix for that. It adds all the migrated CSS files so far to the
debug grd files.

Bug: 1106746
Change-Id: I9b75581e59ecd3290b4fd4eecf1ddbcd46954e18
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2979698
Commit-Queue: Kriti Sapra <kritisapra@google.com>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>