[ts] ExtensionAPI: replace api enum with a const enum
The API enum defines the event and command names used on the messages
sent between the extension server (which is hosted inside devtools) and
the extension API (which is injected into the extension code). This CL
replaces the old object enum with a const enum.

Bug: 1219420
Change-Id: I2dcbdd64f9a80cf7e08962f6ee3447508507deb7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2960610
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>