[Accessibility] Announce selection when accepting suggestions
This CL replaces the anouncement made when accepting a suggestion from a
suggestion box. Instead of announcing the title of the suggestion and
the position in the suggestion list, the screen reader will now announce
the title of the suggestion and that the suggestion has been selected.

Bug: 1223227
Change-Id: Ie9a6e515db76437cb7655cb95ddb3d9c8ca460e2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2984362
Commit-Queue: Olivia Flynn <flynnolivia@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>