[css] Add a separate css class for appending the LMI memory icon
Previously the inline-flex that was applied to align the LMI memory
icon caused the overflow of text not to work correctly for previews.
This adds and uses a separate css class that is applied if we want
to append an LMI memory icon.

Fixed: chromium:1202377
Change-Id: I749aa360792bd3aef39e1d394acce8bda145fd14
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2988313
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>