Delete duplicate Events enum
When porting the CPU throttling functionality from the mobile panel to
the SDK module, the ThrottlingManager.Events enum was added to
the new SDK file but it was not removed from its original location. This
This CL removes the original enum and refactors the dependencies
accordingly.

Bug: none
Change-Id: Ib51239116f01f05678f9f0427f4819469cc87860
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2996749
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>