Demystify _DispatcherPrototype in InspectorBackend
The class _DispatcherPrototype was used in an unusual way to basically
only share the field eventArgs among a couple of class instances. This
CL refactors the code to accomplish the same thing with a standard
pattern: eventArgs is owned by InspectorBackend and shared via a
constructor argument with the new DispatcherManager class instances.

Bug: chromium:1172300
Change-Id: Ia9f76fd55eef12d66f0686c8bfad0bfcb0326ce6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3001851
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>