Remove category from Adorner
In the recent past, many panels have started to use adorners.
Examples are the issues panel and the application panel. The
adorner category, however, is special to the Adorner use-case
in the elements panel.

This CL removed the category property from the Adorner
component and manages categories in the AdornerManager
of the Elements panel.

Bug: chromium:1225910
Change-Id: I18a84f5c5e4668e3dadda405f7349d6455e63b30
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3009821
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>