Separate component for Permissions Policy Section
By creating a separate component for the Permissions Policy Section we
avoid the whole FrameDetailsView rerendering when the user clicks the
'Show Details' button and thus also avoiding the scroll position jumping
around when showing/hiding details.

FrameDetailsView also keeps track of whether the permissions policy
section is expanded or not, such that in case of a re-render this does
not get lost.

Bug: 1218803
Change-Id: Ib676b822be671469a4a08dd0f37307014930feb6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2955361
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>