[Search][Sources] Jump to correct column when revealing search result
This CL adds an optional column number to the SearchMatch class so that
we can move the text cursor to the correct spot when a search result is
revealed. This causes the search term to be shown when pretty printing,
which addresses the pain point described in the bug.
Before/after GIFs: https://imgur.com/a/bahmM8V

Bug: 1227533
Change-Id: I4c00267f9b534d0063b29d0ae8cfd0c83a42a376
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3016612
Commit-Queue: Olivia Flynn <flynnolivia@microsoft.com>
Reviewed-by: Simon Zünd <szuend@chromium.org>