[Debugger] Reschedule update in Debugger if setting breakpoint failed
If we set breakpoints too early (before the back-end is ready), we
receive a ServerError and remove that breakpoint on all targets,
which leads to "breakpoints being flakily removed".

In order to avoid this scenario, we are now rescheduling an
update to the debugger if we do not get a confirmation on
that breakpoint by the back-end, and if we were currently in a
state where we kicked off `Debugger.enable`, but the response
hasn't arrived yet (i.e. Debugger.isReadyToPause == false).

This is only a quickfix to unblock users. The reason for why we do
not just correctly wait for the `Debugger.enable` response
before setting breakpoints, is that other races that reflect in
unreliable breakpoint behavior become more apparent, which has to be
investigated further.

Also-by: bmeurer@chromium.org
Bug: b/190756091, chromium:1229541
Change-Id: If141848f5e6c4a064d324ecde2b2795b99a5406e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3030761
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>