Use chokidar as the file watcher
DISABLE_THIRD_PARTY_CHECK=tvanderlippe

Our file watcher that was rebuilding devtools on changes relied on a
recursive option. The recursive option was only supported on macOS and
Windows, and an ERR_FEATURE_UNAVAILABLE_ON_PLATFORM exception was be
thrown when the option is used on a platform that does not support it.

One of those unsupported platforms was linux, which happens to be a
favorite around these places.

This CL switches to a cross-platform file watcher instead.

Bug: 1229622
Change-Id: Id38d1568999cdff8f054a1e716ee2525fed62fe7
Signed-off-by: Victor Porof <victorporof@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3030767
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>