Highlight location header for attribution reporting API redirects
This CL adds a badge and details text to the "Location" response
header when an attribution for the Attribution Reporting API is
detected.

This information is not reported by the backend, but instead
calculated by the frontend.

Screenshot: https://imgur.com/a/ASyPIae

R=sigurds@chromium.org

Bug: chromium:1190735
Change-Id: Ib62f88e29095f6f174585457a85bd3416edc530c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3023627
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>