Replace custom tooltip implementation with native HTML tooltips
DevTools has had a custom tooltip implementation for a long time.
The primary reason for adopting a custom tooltip implementation
is the lack of styling of a native HTML tooltip.

Unfortunately, maintaining a custom tooltip implementation is
complicated and we regularly run into complicated bugs such as
crbug.com/1219947, crbug.com/1081140 and crbug.com/1217978

Besides that, the remote device view uses the tooltips which
ends up not working, since the remote device view is in a
different frame, which doesn't have access to the custom
tooltip.

All in all, the problems the custom tooltip implementation are
causing are not outweighing the benefits of the slightly
improved styling of the tooltips. The native HTML tooltips
are sufficient for DevTools and adopting the tooltips prevents
a vast variety or problems for our users.

In a follow-up CL, all usages of `install` will be migrating to
setting the `title` attribute directly instead.

R=mraz@chromium.org,jacktfranklin@chromium.org

Bug: 1223391
Change-Id: I06f9317a86dd1c7cc10e52e170b9c62d787f02bd
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3008794
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>