[issues] Do not translate only part of an issue description
This CL removes the translation for 'a secure' and 'an insecure' from
SameSite cookie issues. Given that Markdown issue descriptions are
not yet translated, we should not substitute translated phrases
into the middle of an english issue description.

The translated phrases are still around and will be used once
issue descriptions are localized as well.

R=sigurds@chromium.org

Bug: chromium:1168438
Change-Id: I44035ee58d7df3ceb7a2cc611776c8db49ad8455
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3041612
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>