Do not use REPL mode for live expressions
A small UX improvement for the scenario where:
  1) Pause on exception is enabled
  2) Live expression that throws an exception is used

REPL mode causes the expression to be evaluated as part of the
microtask queue, which is outside the try/catch of Runtime.evaluate.
As REPL mode is not that important for live expressions, we disable
it.

R=yangguo@chromium.org

Fixed: 1225042
Change-Id: Idee4989cf248171baa4fc7e8d103d92dd7c6ffff
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3043609
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>