Updated scripts to resolve paths for css files not in same directory and fixed duplicate entries in `devtools_grd_files`.
Some files import CSS files not in the same directory as them. The
script was currently unable to handle this case and so migrations led
to `File Not Found` errors for the `.css.js` files. The script now
finds the relative path from the imported file to the current file and
adapts the import statement correctly.

Duplicates were being added to `devtools_grd_files` which led to errors
during the build. The check now verifies the entire file path is
contained in the GRD file to prevent duplication.

Bug: 1106746
Change-Id: I5da808a8885bc18477cfa31801a904e17924caab
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3059012
Commit-Queue: Kriti Sapra <kritisapra@google.com>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>