Add tests for SourceFormatter
This requires us using the environment, as the SourceFormatter relies
on numerous singletons related to workspaces, projects and documents.

This CL gets us to about 50% code coverage. The missing coverage is
mostly in the two mapping classes at the end of the file, which are
more difficult to test. We will do that in a follow-up CL.

R=szuend@chromium.org

Bug: 1009443
Change-Id: I8412b5f6e8053d72cea5110a3b82458388fb9e95
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3056378
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>