[l10n] Add remaining locales
The initial locales list was adapted from Lighthouse, which uses only
a subset of the languages available in Chrome. DevTools on the other
hand uses the same set of languages as Chrome. This CL adds the
remainder of the locales to get parity with Chrome, in time for M94
where the localization experiment will be enabled by default.

R=tvanderlippe@chromium.org

Bug: 1163928
Change-Id: Iece95344f1aae7432f51710fe22c6ef667fa5e71
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3069408
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>