Defer extension initialization to after DevTools is done loading
It appears that there is a race between loading DevTools extensions and
recieving the id of the inspected tab. The former is done asynchronously
after DevTools is loaded in the main entrypoint, the latter is sent from
the backend after the DevTools window is loaded. To synchronize
extension loading with the arrival of the inspected tab's ID, this CL
makes sure that the tab ID is there before initialization extensions,
which will queue up until DevTools is ready.

Bug: 1091043
Change-Id: Ieefec0fc2b684c6e0e34f174c5a9913229229936
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3071226
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>