Refactor source map unit test
This introduces a source map helper for tests for encoding source maps
from a human readable form. This is in preparation for changing the
SourceMap class to provide backwards location translation from source
position to a range of locations.  The translation (and ultimately
the test) is needed for fixing stepping with source maps.

Bug: crbug.com/1232347
Change-Id: I6e873fd2d1a714d9db964505dffdb1f47bf01fd8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3087361
Commit-Queue: Jaroslav Sevcik <jarin@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>