Fix flakey memory tests
I know that upping the timeout isn't always the most reliable fix but in
this case locally I was able to replicate the failures until I upped the
timeout this high. The UI that these tests cover is slow and has a lot
of data involved so I'm not surprised that they take such a long time on
the bots.

Bug: 1239550
Change-Id: I62a8065e8d66da9190554cfec3a89a34790babf7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3099990
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>