Styles: fix A11y issue for Toggleing style properties in hosted mode
This CL fixes an A11y bug that causes focus to be lost when toggling
a property in the styles pane using the context menu in hosted mode.
In hosted mode, toggling a property causes focus to be lost.
Moving the reference to the current focused section outside the async
callback causes focus to be retained, and does not affect the inbox
behavior.

Bug: 1239647
Change-Id: I13de55ac4d2292cd970c16f9410966b2e6bfe279
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3094026
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>